move token handler to init app
