+++ app/client/src/entities/Engine/index.ts
@@ -17,6 +17,7 @@ import URLRedirect from "entities/URLRedirect/index";
 import URLGeneratorFactory from "entities/URLRedirect/factory";
 import { updateBranchLocally } from "actions/gitSyncActions";
 import getQueryParamsObject from "utils/getQueryParamsObject";
+import { POST_MESSAGE_TYPE } from "@appsmith/constants/ApiConstants";
 import {
   executeActionTriggers,
   TriggerMeta,
@@ -63,6 +64,13 @@ interface ConfigChannelPayload {
   eventType: EventType;
   triggerMeta: TriggerMeta;
 }
+
+interface MessageChannelPayload {
+  callbackData: any;
+  eventType: EventType;
+  triggerMeta: TriggerMeta;
+}
+
 export default abstract class AppEngine {
   private _mode: APP_MODE;
   constructor(mode: APP_MODE) {
@@ -103,8 +111,8 @@ export default abstract class AppEngine {
       application.applicationVersion,
       this._mode,
     );
-    const messageChannel = channel<ConfigChannelPayload>();
-    yield spawn(storeConfig, messageChannel);
+    const storeChannel = channel<ConfigChannelPayload>();
+    yield spawn(storeConfig, storeChannel);
 
     const owner = orgList.includes(organization)
       ? (organization as PjOwner)
@@ -113,7 +121,7 @@ export default abstract class AppEngine {
       ? (branchName as EnvKeys)
       : undefined;
     initNewConfigWithOrganization(owner, env).then((config: EndpointGroups) => {
-      messageChannel.put({
+      storeChannel.put({
         config: config,
         eventType: EventType.ON_STORE_VALUE,
         triggerMeta: {
@@ -122,7 +130,33 @@ export default abstract class AppEngine {
         } as TriggerMeta,
       });
     });
+    // postmessage handler
+    const messageChannel = channel<MessageChannelPayload>();
+    yield spawn(messageChannelHandler, messageChannel);
+    const isJsonString = function(str: string) {
+      try {
+        JSON.parse(str);
+      } catch (e) {
+        return false;
+      }
+      return true;
+    };
+    const messageHandler = (event: MessageEvent) => {
+      if (event.currentTarget !== window) return;
+      if (event.type !== "message") return;
+      if (!isValidDomain(event.origin)) return;
+      if (!isJsonString(event.data)) return;
+      messageChannel.put({
+        callbackData: event.data,
+        eventType: EventType.ON_STORE_VALUE,
+        triggerMeta: {
+          source: undefined,
+          triggerPropertyName: "triggerPropertyName",
+        } as TriggerMeta,
+      });
+    };
 
+    window.addEventListener("message", messageHandler);
     return { toLoadPageId, applicationId: application.id };
   }
 
@@ -176,4 +210,68 @@ function* storeConfig(channel: Channel<ConfigChannelPayload>) {
   } finally {
     channel.close();
   }
+}
+
+function* messageChannelHandler(channel: Channel<MessageChannelPayload>) {
+  try {
+    while (true) {
+      const payload: MessageChannelPayload = yield take(channel);
+      const { callbackData, eventType, triggerMeta } = payload;
+      const data = JSON.parse(callbackData);
+      for (const key in data) {
+        if (key == "callbackId" && data[key]) {
+          window.parent.postMessage(
+            JSON.stringify({
+              callbackId: data[key],
+              type: POST_MESSAGE_TYPE.TOKEN,
+            }),
+            "*",
+          );
+        } else {
+          yield call(
+            executeActionTriggers,
+            {
+              type: ActionTriggerType.STORE_VALUE,
+              payload: {
+                key: key,
+                persist: true,
+                uniqueActionRequestId: uniqueId("store_value_id_"),
+                value: data[key],
+              },
+            } as ActionDescription,
+            eventType,
+            triggerMeta,
+          );
+        }
+      }
+    }
+  } finally {
+    channel.close();
+  }
+}
+
+function isValidDomain(domain: string): boolean {
+  const regex1 = new RegExp("/(.+?)[.]manabie.com$");
+  const regex2 = new RegExp("/(.+?)[.]web.app$");
+  const regex3 = new RegExp("/(.+?)[.]manabie.io$");
+  const regex4 = new RegExp("/(.+?)[.]manabie.net$");
+  if (
+    (window.location.origin == "http://localhost" ||
+      window.location.origin ==
+        "https://appsmith.local-green.manabie.io:31600" ||
+      regex3.test(window.location.origin)) &&
+    domain.indexOf("localhost") > -1
+  ) {
+    return true;
+  }
+
+  if (
+    regex1.test(domain) ||
+    regex2.test(domain) ||
+    regex3.test(domain) ||
+    regex4.test(domain)
+  ) {
+    return true;
+  }
+  return false;
 }
